pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64", "setuptools_scm>=8"]

[project]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering"
]
dependencies = [
  "cacholote",
  "cads-adaptors@git+https://github.com/ecmwf-projects/cads-adaptors.git",
  "cads-broker@git+https://github.com/ecmwf-projects/cads-broker.git",
  "distributed",
  "fsspec",
  "pydantic",
  "pyyaml",
  "structlog",
  "typer"
]
description = "Utility functions for CADS Dask worker"
dynamic = ["version"]
license = {file = "LICENSE"}
name = "cads-worker"
readme = "README.md"

[project.entry-points."fsspec.specs"]
cci1 = "cads_worker.filesystems:CCI1FileSystem"
cci2 = "cads_worker.filesystems:CCI2FileSystem"

[project.scripts]
cache-cleaner = "cads_worker.entry_points:cache_cleaner"
expire-cache-entries = "cads_worker.entry_points:expire_cache_entries"
init-buckets = "cads_worker.entry_points:init_buckets"

[tool.coverage.run]
branch = true

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "cads_adaptors.*",
  "fsspec.*"
]

[[tool.mypy.overrides]]
ignore_errors = true
module = "cads_worker.filesystems"

[tool.ruff]
# Same as Black.
indent-width = 4
line-length = 88

[tool.ruff.lint]
ignore = [
  # pydocstyle: Missing Docstrings
  "D1"
]
select = [
  # pyflakes
  "F",
  # pycodestyle
  "E",
  "W",
  # isort
  "I",
  # pydocstyle
  "D"
]

[tool.ruff.lint.pycodestyle]
max-line-length = 110

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.setuptools]
packages = ["cads_worker"]

[tool.setuptools_scm]
fallback_version = "999"
write_to = "cads_worker/version.py"
write_to_template = '''
# Do not change! Do not track in version control!
__version__ = "{version}"
'''
